separate lualib from rest of library for app engine
