+++ .gitignore
@@ -7,3 +7,4 @@ tools/debug/
 tmp/
 .DS_Store
 TODO
+.idea
